+++ themes/helium/wordpress/author.php
@@ -11,7 +11,10 @@
 
 defined('ABSPATH') or die;
 
+use Gantry\Framework\Gantry;
+use Gantry\Framework\Theme;
 use Timber\Timber;
+use Timber\User;
 
 /*
  * The template for displaying Author Archive pages
@@ -19,9 +22,9 @@ use Timber\Timber;
 
 global $wp_query;
 
-$gantry = Gantry\Framework\Gantry::instance();
+$gantry = Gantry::instance();
 
-/** @var \Gantry\Framework\Theme $theme */
+/** @var Theme $theme */
 $theme  = $gantry['theme'];
 
 // We need to render contents of <head> before plugin content gets added.
@@ -31,9 +34,9 @@ $context['page_head'] = $theme->render('partials/page_head.html.twig', $context)
 $context['posts'] = Timber::get_posts();
 
 if (isset($authordata)) {
-    $author            = new \Timber\User($authordata->ID);
+    $author            = new User($authordata->ID);
     $context['author'] = $author;
-    $context['title']  = \__('Author:', 'g5_helium') . ' ' . $author->name();
+    $context['title']  = __('Author:', 'g5_helium') . ' ' . $author->name();
 }
 
 Timber::render(['author.html.twig', 'archive.html.twig', 'index.html.twig'], $context);