+++ themes/helium/wordpress/comments.php
@@ -11,6 +11,7 @@
 
 defined('ABSPATH') or die;
 
+use Timber\Post;
 use Timber\Timber;
 
 /*
@@ -19,7 +20,7 @@ use Timber\Timber;
 
 $context = Timber::get_context();
 
-$post            = new \Timber\Post();
+$post            = new Post();
 $context['post'] = $post;
 
 if (post_password_required($post)) {