+++ themes/helium/wordpress/includes/gantry.php
@@ -11,6 +11,7 @@
 
 defined('ABSPATH') or die;
 
+use Gantry5\Loader;
 use Gantry\Framework\Gantry;
 
 try {
@@ -19,16 +20,16 @@ try {
         throw new LogicException('Gantry 5 Framework not found!');
     }
 
-    Gantry5\Loader::setup();
+    Loader::setup();
 
     // Get Gantry instance.
     $gantry = Gantry::instance();
 
     // Initialize the template if not done already.
     if (!isset($gantry['theme.name'])) {
-        $gantry['theme.path'] = \get_stylesheet_directory();
-        $gantry['theme.parent'] = \get_option('template');
-        $gantry['theme.name'] = \get_option('stylesheet');
+        $gantry['theme.path'] = get_stylesheet_directory();
+        $gantry['theme.parent'] = get_option('template');
+        $gantry['theme.name'] = get_option('stylesheet');
     }
 
     // Only a single template can be loaded at any time.
@@ -43,20 +44,20 @@ try {
 
 } catch (Exception $e) {
     // Oops, something went wrong!
-    if (\is_admin()) {
+    if (is_admin()) {
         // In admin display an useful error.
-        \add_action('admin_notices', static function () use ($e) {
+        add_action('admin_notices', static function () use ($e) {
             echo '<div class="error"><p>Failed to load theme: ' . $e->getMessage() . '</p></div>';
         });
         return;
     }
 
-    \add_filter('template_include', static function () {
-        if (\is_customize_preview() && !class_exists('Timber')) {
-            \_e('Timber library plugin not found. ', 'g5_helium');
+    add_filter('template_include', static function () {
+        if (is_customize_preview() && !class_exists('Timber')) {
+            _e('Timber library plugin not found. ', 'g5_helium');
         }
 
-        \_e('Theme cannot be used. For more information, please see the notice in administration.', 'g5_helium');
+        _e('Theme cannot be used. For more information, please see the notice in administration.', 'g5_helium');
 
         die();
     });
@@ -65,12 +66,12 @@ try {
 }
 
 // Hook into administration.
-if (\is_admin()) {
+if (is_admin()) {
     if (file_exists($gantry['theme.path'] . '/admin/init.php')) {
         define('GANTRYADMIN_PATH', $gantry['theme.path'] . '/admin');
     }
 
-    \add_action('init', static function () {
+    add_action('init', static function () {
         if (defined('GANTRYADMIN_PATH')) {
             require_once GANTRYADMIN_PATH . '/init.php';
         }