+++ themes/helium/wordpress/includes/helper.php
@@ -9,23 +9,25 @@
  * http://www.gnu.org/licenses/gpl-2.0.html
  */
 
+use Timber\Timber;
+
 defined('ABSPATH') or die;
 
 // Extend Timber context
-\add_filter('timber_context', array('G5ThemeHelper', 'add_to_context'));
+add_filter('timber_context', array('G5ThemeHelper', 'add_to_context'));
 
 // Modify the default Admin Bar margins to render properly in the mobile mode
-\add_theme_support('admin-bar', array('callback' => array('G5ThemeHelper', 'admin_bar_margins')));
+add_theme_support('admin-bar', array('callback' => array('G5ThemeHelper', 'admin_bar_margins')));
 
 // Add comments pagination link attributes
-\add_filter('previous_comments_link_attributes', array('G5ThemeHelper', 'comments_pagination_attributes'));
-\add_filter('next_comments_link_attributes', array('G5ThemeHelper', 'comments_pagination_attributes'));
+add_filter('previous_comments_link_attributes', array('G5ThemeHelper', 'comments_pagination_attributes'));
+add_filter('next_comments_link_attributes', array('G5ThemeHelper', 'comments_pagination_attributes'));
 
 // Change single post pagination to list items
-\add_filter('wp_link_pages_link', array('G5ThemeHelper', 'wp_link_pages_li_return'));
+add_filter('wp_link_pages_link', array('G5ThemeHelper', 'wp_link_pages_li_return'));
 
 // Modify Tag Cloud widget arguments
-\add_filter('widget_tag_cloud_args', array('G5ThemeHelper', 'tag_cloud_widget_modified_args'));
+add_filter('widget_tag_cloud_args', array('G5ThemeHelper', 'tag_cloud_widget_modified_args'));
 
 /**
  * Helper class G5ThemeHelper containing useful theme functions and hooks
@@ -40,8 +42,8 @@ class G5ThemeHelper
      */
     public static function add_to_context(array $context)
     {
-        $context['is_user_logged_in'] = \is_user_logged_in();
-        $context['pagination']        = Timber\Timber::get_pagination();
+        $context['is_user_logged_in'] = is_user_logged_in();
+        $context['pagination']        = Timber::get_pagination();
 
         return $context;
     }
@@ -59,36 +61,36 @@ class G5ThemeHelper
     {
         $GLOBALS['comment'] = $comment; ?>
 
-        <li id="comment-<?php \comment_ID(); ?>" <?php \comment_class(); ?>>
-        <article id="div-comment-<?php \comment_ID(); ?>" class="comment-body">
+        <li id="comment-<?php comment_ID(); ?>" <?php comment_class(); ?>>
+        <article id="div-comment-<?php comment_ID(); ?>" class="comment-body">
             <span class="child-arrow-indicator"><i class="fa fa-arrow-up" aria-hidden="true"></i></span>
             <header class="comment-author">
                 <div class="author-avatar">
-                    <?php echo \get_avatar($comment, $size = '40'); ?>
+                    <?php echo get_avatar($comment, $size = '40'); ?>
                 </div>
                 <div class="author-meta vcard">
-                    <?php printf(\__('<span class="author-name">%s</span>', 'g5_helium'), \get_comment_author_link()); ?>
+                    <?php printf(__('<span class="author-name">%s</span>', 'g5_helium'), get_comment_author_link()); ?>
                     <br />
-                    <time datetime="<?php echo \comment_date('c'); ?>">
-                        <a href="<?php echo \esc_url(\get_comment_link($comment->comment_ID)); ?>">
-                            <?php printf(\__('Commented on %1$s', 'g5_helium'), \get_comment_date(), \get_comment_time()); ?>
+                    <time datetime="<?php echo comment_date('c'); ?>">
+                        <a href="<?php echo esc_url(get_comment_link($comment->comment_ID)); ?>">
+                            <?php printf(__('Commented on %1$s', 'g5_helium'), get_comment_date(), get_comment_time()); ?>
                         </a>
                     </time>
-                    <?php \edit_comment_link(\__('(Edit)', 'g5_helium'), '<span class="edit-link">', '</span>'); ?>
+                    <?php edit_comment_link(__('(Edit)', 'g5_helium'), '<span class="edit-link">', '</span>'); ?>
                 </div>
             </header>
 
             <section class="comment-content">
                 <?php if ($comment->comment_approved == '0') : ?>
                     <div class="notice">
-                        <p class="alert-info"><?php \_e('Your comment is awaiting moderation.', 'g5_helium'); ?></p>
+                        <p class="alert-info"><?php _e('Your comment is awaiting moderation.', 'g5_helium'); ?></p>
                     </div>
                 <?php endif; ?>
 
-                <?php \comment_text(); ?>
+                <?php comment_text(); ?>
             </section>
 
-            <?php \comment_reply_link(array_merge($args,
+            <?php comment_reply_link(array_merge($args,
                 array('add_below' => 'div-comment', 'before' => '<div class="comment-reply">', 'after' => '</div>', 'depth' => $depth, 'max_depth' => $args['max_depth']))); ?>
 
         </article>
@@ -133,7 +135,7 @@ class G5ThemeHelper
             'order'    => 'DESC'
         );
 
-        return \wp_parse_args($new_args, $args);
+        return wp_parse_args($new_args, $args);
     }
 
     /**