+++ themes/hydrogen/wordpress/buddypress.php
@@ -11,15 +11,18 @@
 
 defined('ABSPATH') or die;
 
+use Gantry\Framework\Gantry;
+use Gantry\Framework\Theme;
+use Timber\Helper;
 use Timber\Timber;
 
 /*
  * The template for displaying BBPress pages
  */
 
-$gantry = Gantry\Framework\Gantry::instance();
+$gantry = Gantry::instance();
 
-/** @var \Gantry\Framework\Theme $theme */
+/** @var Theme $theme */
 $theme  = $gantry['theme'];
 
 // We need to render contents of <head> before plugin content gets added.
@@ -27,6 +30,6 @@ $context              = Timber::get_context();
 $context['page_head'] = $theme->render('partials/page_head.html.twig', $context);
 
 $context['posts']   = Timber::query_post();
-$context['content'] = \Timber\Helper::ob_function('the_content');
+$context['content'] = Helper::ob_function('the_content');
 
 Timber::render('bbpress.html.twig', $context);