+++ themes/hydrogen/wordpress/comments.php
@@ -11,6 +11,7 @@
 
 defined('ABSPATH') or die;
 
+use Timber\Post;
 use Timber\Timber;
 
 /*
@@ -19,10 +20,10 @@ use Timber\Timber;
 
 $context = Timber::get_context();
 
-$post            = new \Timber\Post();
+$post            = new Post();
 $context['post'] = $post;
 
-if (\post_password_required($post)) {
+if (post_password_required($post)) {
     return;
 }