+++ themes/hydrogen/wordpress/footer.php
@@ -21,11 +21,10 @@ use Timber\Timber;
 $timberContext = $GLOBALS['timberContext'];
 
 if (!isset($timberContext)) {
-    throw new \Exception('Timber context not set in footer.');
+    throw new RuntimeException('Timber context not set in footer.');
 }
 
-$timberContext['content'] = ob_get_contents();
-ob_end_clean();
+$timberContext['content'] = ob_get_clean();
 
 $templates = ['page-plugin.html.twig'];
 Timber::render($templates, $timberContext);