+++ themes/hydrogen/wordpress/search.php
@@ -11,22 +11,24 @@
 
 defined('ABSPATH') or die;
 
+use Gantry\Framework\Gantry;
+use Gantry\Framework\Theme;
 use Timber\Timber;
 
 /*
  * Search results page
  */
 
-$gantry = Gantry\Framework\Gantry::instance();
+$gantry = Gantry::instance();
 
-/** @var \Gantry\Framework\Theme $theme */
+/** @var Theme $theme */
 $theme  = $gantry['theme'];
 
 // We need to render contents of <head> before plugin content gets added.
 $context              = Timber::get_context();
 $context['page_head'] = $theme->render('partials/page_head.html.twig', $context);
 
-$context['title'] = \__('Search results for:', 'g5_hydrogen') . ' ' . \get_search_query();
+$context['title'] = __('Search results for:', 'g5_hydrogen') . ' ' . get_search_query();
 $context['posts'] = Timber::get_posts();
 
 $templates = ['search.html.twig', 'archive.html.twig', 'index.html.twig'];