+++ themes/hydrogen/wordpress/single.php
@@ -11,23 +11,24 @@
 
 defined('ABSPATH') or die;
 
+use Gantry\Framework\Gantry;
+use Gantry\Framework\Theme;
 use Timber\Timber;
 
 /*
  * The Template for displaying all single posts
  */
 
-$gantry = Gantry\Framework\Gantry::instance();
+$gantry = Gantry::instance();
 
-/** @var \Gantry\Framework\Theme $theme */
+/** @var Theme $theme */
 $theme  = $gantry['theme'];
 
 // We need to render contents of <head> before plugin content gets added.
 $context              = Timber::get_context();
 $context['page_head'] = $theme->render('partials/page_head.html.twig', $context);
 
-$post = Timber::query_post();
-
+$post            = Timber::query_post();
 $context['post'] = $post;
 $context['wp_title'] .= ' - ' . $post->title();